requirements.txt
pandas==2.2.2
Pillow==10.3.0
Requests==2.32.3
streamlit==1.35.0
streamlit_autorefresh==1.0.1
streamlit_option_menu==0.3.13
